refactor: split stats collectors
